make Attributes required, add str device_class to AttributesBase, move simple states into simple.py, move sensor literals into const
