+++ src/hassette/models/states/binary_sensor.py
@@ -1,14 +1,12 @@
 from typing import Literal
 
-from pydantic import Field, field_validator
+from pydantic import field_validator
 
-from hassette.models.states.base import AttributesBase, BaseState
+from hassette.models.states.base import BaseState
 
 
 class BinarySensorState(BaseState[bool | None]):
     domain: Literal["binary_sensor"]
-
-    attributes: AttributesBase | None = Field(default=None)
 
     @field_validator("value")
     @classmethod
