+++ src/hassette/models/states/camera.py
@@ -8,11 +8,10 @@ from .base import AttributesBase, StringBaseState
 class CameraState(StringBaseState):
     class Attributes(AttributesBase):
         access_token: SecretStr | None = Field(default=None)
-        device_class: str | None = Field(default=None)
         model_name: str | None = Field(default=None)
         brand: str | None = Field(default=None)
         entity_picture: str | None = Field(default=None)
 
     domain: Literal["camera"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
