+++ src/hassette/models/states/climate.py
@@ -25,4 +25,4 @@ class ClimateState(StringBaseState):
 
     domain: Literal["climate"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
