--- src/hassette/models/states/conversation.py
@@ -1,11 +1,0 @@
-from typing import Literal
-
-from pydantic import Field
-
-from .base import AttributesBase, StringBaseState
-
-
-class ConversationState(StringBaseState):
-    domain: Literal["conversation"]
-
-    attributes: AttributesBase | None = Field(default=None)
