+++ src/hassette/models/states/device_tracker.py
@@ -27,4 +27,4 @@ class DeviceTrackerState(StringBaseState):
 
     domain: Literal["device_tracker"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
