+++ src/hassette/models/states/humidifier.py
@@ -17,4 +17,4 @@ class HumidifierState(StringBaseState):
 
     domain: Literal["humidifier"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
