+++ src/hassette/models/states/media_player.py
@@ -8,10 +8,7 @@ from .base import AttributesBase, StringBaseState
 class MediaPlayerState(StringBaseState):
     class Attributes(AttributesBase):
         assumed_state: bool | None = Field(default=None)
-        device_class: str | None = Field(default=None)
         adb_response: Any | None = Field(default=None)
         hdmi_input: Any | None = Field(default=None)
 
     domain: Literal["media_player"]
-
-    attributes: Attributes | None = Field(default=None)
