--- src/hassette/models/states/stt.py
@@ -1,11 +1,0 @@
-from typing import Literal
-
-from pydantic import Field
-
-from .base import AttributesBase, StringBaseState
-
-
-class SttState(StringBaseState):
-    domain: Literal["stt"]
-
-    attributes: AttributesBase | None = Field(default=None)
