+++ src/hassette/models/states/sun.py
@@ -19,4 +19,4 @@ class SunState(StringBaseState):
 
     domain: Literal["sun"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
