--- src/hassette/models/states/tts.py
@@ -1,12 +1,0 @@
-from typing import Literal
-
-from pydantic import Field
-from whenever import Instant
-
-from .base import AttributesBase, BaseState
-
-
-class TtsState(BaseState[Instant | None]):
-    domain: Literal["tts"]
-
-    attributes: AttributesBase | None = Field(default=None)
