+++ src/hassette/models/states/vacuum.py
@@ -15,4 +15,4 @@ class VacuumState(StringBaseState):
 
     domain: Literal["vacuum"]
 
-    attributes: Attributes | None = Field(default=None)
+    attributes: Attributes
